requirements.txt
aiogram
tensorflow
numpy
keras
Pillow